로그인 처리중...

+++ yosi/src/main/java/org/yosi/dto/BoardDTO.java
@@ -13,6 +13,6 @@ public class BoardDTO {
 	private String writer;
 	private Date regdate;
 	private Integer viewcnt;
-	private int replycnt;
+	private Long replycnt;
 
 }
