+++ yosi/src/main/java/org/yosi/dto/ReplyDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Data
 public class ReplyDTO {
 	
-	private Integer rno;
+	private Long rno;
 	private String reply, replyer;
 	private Date replydate;
 	private Long tno;
