+++ yosi/src/main/java/org/yosi/mapper/ReplyMapper.java
@@ -20,7 +20,7 @@ public interface ReplyMapper {
 	public void create(ReplyDTO dto);
 	
 	@Delete("delete from tbl_reply_board where rno = #{rno}")
-	public void delete(Integer rno);
+	public void delete(Long rno);
 	
 	@Update("update tbl_reply_board set reply = #{reply}, replyer = #{replyer} where rno = #{rno}")
 	public void update(ReplyDTO dto);
